Refactor menuoutlineview in another file
